refactor: make more explicit the relationship btwn TableParquetOptions vs ParquetOptions vs WriterProperties
