+++ datafusion/common/src/file_options/parquet_writer.rs
@@ -17,18 +17,13 @@
 
 //! Options related to how parquet files should be written
 
-use crate::{
-    config::{ParquetOptions, TableParquetOptions},
-    DataFusionError, Result,
-};
+use crate::{config::TableParquetOptions, DataFusionError, Result};
 
 use parquet::{
     basic::{BrotliLevel, GzipLevel, ZstdLevel},
-    file::{
-        metadata::KeyValue,
-        properties::{EnabledStatistics, WriterProperties, WriterVersion},
+    file::properties::{
+        EnabledStatistics, WriterProperties, WriterPropertiesBuilder, WriterVersion,
     },
-    schema::types::ColumnPath,
 };
 
 /// Options for writing parquet files
@@ -52,140 +47,11 @@ impl ParquetWriterOptions {
 impl TryFrom<&TableParquetOptions> for ParquetWriterOptions {
     type Error = DataFusionError;
 
-    fn try_from(parquet_options: &TableParquetOptions) -> Result<Self> {
-        let ParquetOptions {
-            data_pagesize_limit,
-            write_batch_size,
-            writer_version,
-            dictionary_page_size_limit,
-            max_row_group_size,
-            created_by,
-            column_index_truncate_length,
-            data_page_row_count_limit,
-            bloom_filter_on_write,
-            encoding,
-            dictionary_enabled,
-            compression,
-            statistics_enabled,
-            max_statistics_size,
-            bloom_filter_fpp,
-            bloom_filter_ndv,
-            //  below is not part of ParquetWriterOptions
-            enable_page_index: _,
-            pruning: _,
-            skip_metadata: _,
-            metadata_size_hint: _,
-            pushdown_filters: _,
-            reorder_filters: _,
-            allow_single_file_parallelism: _,
-            maximum_parallel_row_group_writers: _,
-            maximum_buffered_record_batches_per_stream: _,
-            bloom_filter_on_read: _,
-        } = &parquet_options.global;
-
-        let key_value_metadata = if !parquet_options.key_value_metadata.is_empty() {
-            Some(
-                parquet_options
-                    .key_value_metadata
-                    .clone()
-                    .drain()
-                    .map(|(key, value)| KeyValue { key, value })
-                    .collect::<Vec<_>>(),
-            )
-        } else {
-            None
-        };
-
-        let mut builder = WriterProperties::builder()
-            .set_data_page_size_limit(*data_pagesize_limit)
-            .set_write_batch_size(*write_batch_size)
-            .set_writer_version(parse_version_string(writer_version.as_str())?)
-            .set_dictionary_page_size_limit(*dictionary_page_size_limit)
-            .set_max_row_group_size(*max_row_group_size)
-            .set_created_by(created_by.clone())
-            .set_column_index_truncate_length(*column_index_truncate_length)
-            .set_data_page_row_count_limit(*data_page_row_count_limit)
-            .set_bloom_filter_enabled(*bloom_filter_on_write)
-            .set_key_value_metadata(key_value_metadata);
-
-        if let Some(encoding) = &encoding {
-            builder = builder.set_encoding(parse_encoding_string(encoding)?);
-        }
-
-        if let Some(enabled) = dictionary_enabled {
-            builder = builder.set_dictionary_enabled(*enabled);
-        }
-
-        if let Some(compression) = &compression {
-            builder = builder.set_compression(parse_compression_string(compression)?);
-        }
-
-        if let Some(statistics) = &statistics_enabled {
-            builder =
-                builder.set_statistics_enabled(parse_statistics_string(statistics)?);
-        }
-
-        if let Some(size) = max_statistics_size {
-            builder = builder.set_max_statistics_size(*size);
-        }
-
-        if let Some(fpp) = bloom_filter_fpp {
-            builder = builder.set_bloom_filter_fpp(*fpp);
-        }
-
-        if let Some(ndv) = bloom_filter_ndv {
-            builder = builder.set_bloom_filter_ndv(*ndv);
-        }
-
-        for (column, options) in &parquet_options.column_specific_options {
-            let path = ColumnPath::new(column.split('.').map(|s| s.to_owned()).collect());
-
-            if let Some(bloom_filter_enabled) = options.bloom_filter_enabled {
-                builder = builder
-                    .set_column_bloom_filter_enabled(path.clone(), bloom_filter_enabled);
-            }
-
-            if let Some(encoding) = &options.encoding {
-                let parsed_encoding = parse_encoding_string(encoding)?;
-                builder = builder.set_column_encoding(path.clone(), parsed_encoding);
-            }
-
-            if let Some(dictionary_enabled) = options.dictionary_enabled {
-                builder = builder
-                    .set_column_dictionary_enabled(path.clone(), dictionary_enabled);
-            }
-
-            if let Some(compression) = &options.compression {
-                let parsed_compression = parse_compression_string(compression)?;
-                builder =
-                    builder.set_column_compression(path.clone(), parsed_compression);
-            }
-
-            if let Some(statistics_enabled) = &options.statistics_enabled {
-                let parsed_value = parse_statistics_string(statistics_enabled)?;
-                builder =
-                    builder.set_column_statistics_enabled(path.clone(), parsed_value);
-            }
-
-            if let Some(bloom_filter_fpp) = options.bloom_filter_fpp {
-                builder =
-                    builder.set_column_bloom_filter_fpp(path.clone(), bloom_filter_fpp);
-            }
-
-            if let Some(bloom_filter_ndv) = options.bloom_filter_ndv {
-                builder =
-                    builder.set_column_bloom_filter_ndv(path.clone(), bloom_filter_ndv);
-            }
-
-            if let Some(max_statistics_size) = options.max_statistics_size {
-                builder =
-                    builder.set_column_max_statistics_size(path, max_statistics_size);
-            }
-        }
-
+    fn try_from(parquet_table_options: &TableParquetOptions) -> Result<Self> {
         // ParquetWriterOptions will have defaults for the remaining fields (e.g. sorting_columns)
         Ok(ParquetWriterOptions {
-            writer_options: builder.build(),
+            writer_options: WriterPropertiesBuilder::try_from(parquet_table_options)?
+                .build(),
         })
     }
 }
